show connected cars, rename Controller folder to Controllers -> naming violation
